// src/types/websocket-events.ts
export enum WebsocketServerEvent {
  AUTH_ERROR = 'AUTH_ERROR',
  ONLY_ADMIN_PERMITTED = 'ONLY_ADMIN_PERMITTED',
  CONNECTED_USERS = 'CONNECTED_USERS',
  BLOCK_RESTAURANT_REVIEWS_CHANGED = 'BLOCK_RESTAURANT_REVIEWS_CHANGED',
  INPUT_ERROR = 'INPUT_ERROR',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
}

export enum WebsocketClientEvent {
  CHANGE_BLOCK_RESTAURANT_REVIEWS = 'CHANGE_BLOCK_RESTAURANT_REVIEWS',
}